requirements.txt
Django==5.0
django-environ==0.11.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7  # для PostgreSQL
whitenoise==6.6.0       # для статических файлов